% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.tbl_biogridr}
\alias{aggregate.tbl_biogridr}
\title{Aggregate interaction data from BioGRID}
\usage{
\method{aggregate}{tbl_biogridr}(x, neg = NULL, pos = NULL, phy = NULL,
  und = NULL)
}
\arguments{
\item{x}{An object of class \code{tbl_biogridr} (i.e. interaction data derived 
from the 'interactions' table in your local BioGRID database).}

\item{neg}{Character vector of interaction system names that correspond to 
negative interactions (e.g. 'Negative Genetic').}

\item{pos}{Character vector of interaction system names that correspond to 
positive interactions (e.g. 'Positive Genetic').}

\item{phy}{Character vector of interaction system names that correspond to 
physical interactions (e.g. 'Two-hybrid').}

\item{und}{Character vector of interaction system names that should be 
treated as undirected (see Details).}
}
\description{
Aggregates BioGRID interaction data based on several categories. 
For more information see Details.
}
\details{
Interactions in the BioGRID database are stored in an edge list format 
where each row corresponds to an observed interaction between two gene 
identifiers. Interactions are defined by the experimental system used to 
identify the relationship between two genes. Most of these experimental 
systems can be categorized into genetic interactions, which can be negative 
or positive, and physical interactions. \cr\cr

\bold{Undirected interactions:} \cr
Some experimental systems such as 'Synthetic Lethality' (SL) involve the 
deletion of either interacting gene, which makes them undirected since 
\code{A_mutant <-SL-> B_mutant} is equivalent to 
\code{B_mutant <-SL-> A_mutant}. For this reason, the negative interaction 
\code{A -> B} can be aggregated with the negative interaction \code{B -> A}. 
In this case the resulting table will include a row for \code{A -> B} and 
\code{B -> A}. Interaction systems such as 'Synthetic Dosage Lethality' are 
directed because \code{A_overexpressed <-SDL-> B_mutant} is not equivalent to
\code{B_overexpressed <-SDL-> A_mutant}. By default, such directed 
interaction systems are aggregated into the category 'other'. For
growth based experimental systems nodes would ideally be based on 
perturbation rather than gene, which would allow one to easily aggregate 
\code{A_perturbation <-neg/pos-> B_purturbation} with 
\code{B_perturbation <-neg/pos-> A_perturbation}.
}
\examples{
\dontrun{
src_biogridr() \%>\%
  outer_net('CTF4') \%>\%
  aggregate
}

}
