% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Cluster rows or columns of a matrix}
\usage{
cluster(m, method = "ward.D2", correlation = "pearson",
  metric = "euclidean", rows = T, cols = T)
}
\arguments{
\item{m}{Numeric matrix.}

\item{method}{Clustering method passed to \link[stats]{hclust}. Defaults to
'complete'.}

\item{rows}{Logical. Whether to cluster rows. Defaults to TRUE.}

\item{cols}{Logical. Whether to cluster columns. Defaults to TRUE.}
}
\value{
Matrix with columns and/or rows sorted.
}
\description{
Convenience function for sorting the rows and/or columns of a 
matrix based on \link[stats]{hclust}.

Given a numeric matrix, or data frame, sort rows and/or 
columns based on correlation distance heirarchical clustering.
}

