% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtile.R
\name{ggtile}
\alias{ggtile}
\title{Plot a numeric matrix}
\usage{
ggtile(m, diverging = TRUE, ...)
}
\arguments{
\item{m}{A numeric matrix}

\item{diverging}{Logical. Whether to use a diverging color scale.}

\item{...}{other arguments passed to scale_fill_gradient 
\link[ggplot2]{discrete_scale}.}
}
\value{
A ggplot object.
}
\description{
Convenience function for plotting a numeric matrix as colored 
tiles using ggplot with row and column order preserved. Combine with 
\link{cluster} for a clustered heatmap.
}

