% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_biogrid.R
\name{initialize_biogrid}
\alias{initialize_biogrid}
\title{Initialize Local BioGRID Database}
\usage{
initialize_biogrid(db = getOption("biogrid.db"),
  url = getOption("biogrid.url"))
}
\arguments{
\item{db}{Path to an SQLite database. Defaults to
\code{getOption('biogrid.db')}.}

\item{url}{URL to BIOGRID-ALL-<VERSION>.tab2.zip archive. Defaults to
\code{getOption('biogrid.url')}.}
}
\description{
Initialize your local BioGRID database with the latest interactions, or
previously archived versions of the BioGRID database.
}
\details{
If the local BioGRID database has not been initialized, this function will
run \link{update_biogrid} which downloads a BIOGRID-ALL-<VERSION>.tab2.zip
to a temporary file, which is then written to an SQLite database.
}
\seealso{
\link{update_biogrid}
}

