% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inner_net.R
\name{inner_net}
\alias{inner_net}
\alias{inner_net.src_biogridr}
\title{Select edges connecting a set of nodes}
\usage{
inner_net(src, ...)

\method{inner_net}{src_biogridr}(src, genes, org = organism("cerevisiae"),
  ...)
}
\arguments{
\item{src}{An interaction network source (e.g. the result of \code{src_biogridr()}).}

\item{...}{other arguments passed to methods. For \code{src_biogridr} method,
additional arguments are passed to \link[dplyr]{filter}.}

\item{genes}{Gene identifiers (i.e. nodes).}

\item{org}{NCBI organism IDs used to filter BioGRID interactions. Defaults to
\code{organism('cerevisiae')}.}
}
\description{
This is a generic function for selecting network edges connecting
specified nodes. This "inner" network, where all edges connect only a set of
nodes, can be contrasted to an "outer" network which contains all edges
departing a set of nodes.
}
\examples{
\dontrun{
genes <- c('CTF4', 'TOF1')
src_biogridr() \%>\% inner_net(genes)
src_biogridr() \%>\% outer_net(genes)
}
}
\seealso{
\link{outer_net}
}

