% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organism.R
\name{organism}
\alias{organism}
\title{Search for NCBI organism ID}
\usage{
organism(x, ...)
}
\arguments{
\item{x}{a regular expression used to match genus-species names.}

\item{...}{additional arguments passed to \link[base]{grepl}.}
}
\description{
A convenience function that returns NCBI organism IDs based on a 
genus-species name. This is useful for filtering interactions for specific
organisms.
}
\examples{
\dontrun{
organism('cerevisiae')
}

}

