% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outer_net.R
\name{outer_net}
\alias{outer_net}
\alias{outer_net.src_biogridr}
\title{Select edges departing or arriving at a set of nodes}
\usage{
outer_net(src, ...)

\method{outer_net}{src_biogridr}(src, genes, org = organism("cerevisiae"),
  ...)
}
\arguments{
\item{src}{An interaction network source (e.g. the result of \code{src_biogridr()}).}

\item{...}{other arguments passed to methods. For \code{src_biogridr} method, 
additional arguments are passed to \link[dplyr]{filter}.}

\item{genes}{Gene identifiers (i.e. nodes).}

\item{org}{NCBI organism IDs used to filter BioGRID interactions. Defaults to 
\code{organism('cerevisiae')}.}
}
\description{
This is a generic function for selecting network edges departing
from or arriving at a specific nodes. This "outer" network can be contrasted 
to an "inner" network which contains only edges connecting within a set of 
nodes.
}
\examples{
\dontrun{
genes <- c('CTF4', 'TOF1')
src_biogridr() \%>\% inner_net(genes)
src_biogridr() \%>\% outer_net(genes)
}

}
\seealso{
\link{outer_net}
}
