% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tbl_biogridr_aggregated}
\alias{plot.tbl_biogridr_aggregated}
\title{Plot a heatmap of aggregated BioGRID interactions}
\usage{
\method{plot}{tbl_biogridr_aggregated}(x, type = "score", bounds = c(-3, 3),
  density = 1, clust = TRUE, ...)
}
\arguments{
\item{x}{Aggregated BioGRID interactions (class = tbl_biogridr_aggregated)}

\item{type}{A string indicating which type of interaction to plot ('score', 
'positive', 'negative', 'physical', 'other')}

\item{clust}{Whether or not to cluster the data. Defaults to TRUE.}

\item{...}{additional arguments passed to \link{cluster}.}

\item{trunc}{value to truncate data. Truncation occurs before clustering. 
Defaults to 3.}
}
\description{
Plot a clustered heatmap of aggregated biogrid interactions.
}

