% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/src_biogrid.R
\name{src_biogrid}
\alias{biogrid}
\alias{src_biogrid}
\title{Connect to local BioGRID database}
\usage{
src_biogrid(path = getOption("biogrid.db"))
}
\arguments{
\item{path}{Path to BioGRID SQLite database. Defaults to
\code{getOption('biogrid.db')}.}
}
\description{
Connect to the biogrid package's interaction database
}
\details{
BioGRID interactions are stored in a local SQLite database, and
\code{src_biogrid()} returns a connection to this database via
\link[dplyr]{dplyr}. This allows the user to aviod loading all 700,000+
interactions into memory by using \code{dplyr} to \link[dplyr]{select}
columns and \link[dplyr]{filter} rows.
}
\examples{
\dontrun{

# First download all interaction data
# WARNING: This takes a few minutes!
update_biogrid()

# Connect to the database
src_biogrid()

# Query the database for a CTF4 outer network
src_biogrid() \%>\%
  outer_net('CTF4')

# Aggregate the CTF4 outer network
src_biogrid() \%>\%
  outer_net('CTF4') \%>\%
  aggregate

# Use dplyr to make custom queries to the database
library(dplyr)
src_biogrid() \%>\%
  tbl('systems')

src_biogrid() \%>\%
  tbl('organisms')

genes <- c('TOF1', 'CTF4')
src_biogrid() \%>\%
  tbl('interactions') \%>\%
  filter(
    (a \%in\% genes | b \%in\% genes),
     a_organism == organism('cerevisiae'),
     b_organism == organism('cerevisiae')) \%>\%
  select(a, b)
}
}

