% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_biogrid.R
\name{update_biogrid}
\alias{update_biogrid}
\title{Update Local BioGRID Database}
\usage{
update_biogrid(db = getOption("biogrid.db"), url = getOption("biogrid.url"))
}
\arguments{
\item{db}{Path to an SQLite database. Defaults to
\code{getOption('biogrid.db')}.}

\item{url}{URL to BIOGRID-ALL-<VERSION>.tab2.zip archive. Defaults to
\code{getOption('biogrid.url')}.}
}
\description{
Update your local BioGRID database with the latest interactions, or
previously archived versions of the BioGRID database.
}
\details{
This function downloads a BIOGRID-ALL-<VERSION>.tab2.zip to a temporary file,
which is then written to an SQLite database. The default biogrid url
(\code{getOption('biogrid.url')}) is the latest release.
}

