% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_biogrid.R
\name{update_biogridr}
\alias{update_biogridr}
\title{Update Local BioGRID Database}
\usage{
update_biogridr(db = getOption("biogridr.db"),
  url = getOption("biogridr.url"))
}
\arguments{
\item{db}{Path to an SQLite database. Defaults to
\code{getOption('biogridr.db')}.}

\item{url}{URL to BIOGRID-ALL-<VERSION>.tab2.zip archive. Defaults to
\code{getOption('biogridr.url')}.}
}
\description{
Update your local BioGRID database with the latest interactions, or
previously archived versions of the BioGRID database.
}
\details{
This function downloads a BIOGRID-ALL-<VERSION>.tab2.zip to a temporary file,
which is then written to an SQLite database. The default biogridr url
(\code{getOption('biogridr.url')}) is the latest release.
}

