% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/src_biogrid.R
\name{src_biogridr}
\alias{biogridr}
\alias{src_biogridr}
\title{Connect to local BioGRID database}
\usage{
src_biogridr(path = getOption("biogridr.db"))
}
\arguments{
\item{path}{Path to BioGRID SQLite database. Defaults to
\code{getOption('biogridr.db')}.}
}
\description{
Connect to the biogridr package's interaction database
}
\details{
BioGRID interactions are stored in a local SQLite database, and
\code{src_biogridr()} returns a connection to this database via
\link[dplyr]{dplyr}. This allows the user to aviod loading all 700,000+
interactions into memory by using \code{dplyr} to \link[dplyr]{select}
columns and \link[dplyr]{filter} rows.
}
\examples{
\dontrun{

# First download all interaction data
# WARNING: This takes a few minutes!
update_biogridr()

# Connect to the database
src_biogridr()

# Query the database for a CTF4 outer network
src_biogridr() \%>\%
  outer_net('CTF4')

# Aggregate the CTF4 outer network
src_biogridr() \%>\%
  outer_net('CTF4') \%>\%
  aggregate

# Use dplyr to make custom queries to the database
library(dplyr)
src_biogridr() \%>\%
  tbl('systems')

src_biogridr() \%>\%
  tbl('organisms')

genes <- c('TOF1', 'CTF4')
src_biogridr() \%>\%
  tbl('interactions') \%>\%
  filter(
    (a \%in\% genes | b \%in\% genes),
     a_organism == organism('cerevisiae'),
     b_organism == organism('cerevisiae')) \%>\%
  select(a, b)
}
}

